pagination

+++ essay-genius-backend/essay-service/src/main/java/com/phat/api/model/response/EssaySaveResponse.java
@@ -14,5 +14,5 @@ public class EssaySaveResponse {
     String promptText;
     EssayResponseWrapper<EssayTaskTwoScoreResponse> essayTaskTwoScoreResponse;
     String visibility;
-    byte band;
+    Double band;
 }
